Preliminary change for https://github.com/jpluimers/GExperts/issues/4
Calls the original `OnDestroy` value as well as restoring it because the restore point is inside the `OnDestroy` handling.

In addition a few comments now explain what `TManagedForm` is for as it gave me some context in where/how this might be tested.

This might fix the issue completely, but needs thorough testing and code review as this has been broken for a while so I'm not sure if anything relies on the broken behaviour.